Move CheckOffs UI to Classes/CheckOffKids.jsx
